backend/requirements.txt
anyio==4.12.0
asgiref==3.11.0
click==8.3.1
colorama==0.4.6
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httptools==0.7.1
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
